add domain layer for location and weather

+++ app/src/main/java/com/ipro/weatherapp/domain/model/HourlyTemperature.kt
@@ -1,6 +1,6 @@
 package com.ipro.weatherapp.domain.model
 
 data class HourlyTemperature(
-    val temperature: Double,
+    val temperature: Int,
     val hour: Int,
 )
